# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Portal API Documentation",
        "description": "API Documentation untuk Job Portal Application. Sistem ini menyediakan endpoint untuk manajemen job postings, aplikasi pekerjaan, dan autentikasi user.",
        "contact": {
            "name": "Job Portal Support",
            "email": "support@jobportal.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/jobs/{jobId}/apply": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Apply for a job",
                "operationId": "06792d590f8e1daf2359b4bb039973d1",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "cover_letter",
                                    "cv"
                                ],
                                "properties": {
                                    "cover_letter": {
                                        "type": "string",
                                        "example": "I am interested in this position..."
                                    },
                                    "cv": {
                                        "description": "CV file (PDF, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Application submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Application submitted successfully"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get all applications (admin only)",
                "operationId": "895749895f8d7900e2b7a93830d1f67f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "reviewed",
                                "accepted",
                                "rejected"
                            ]
                        }
                    },
                    {
                        "name": "job_id",
                        "in": "query",
                        "description": "Filter by job ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of applications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Application"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/my-applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get current user's applications",
                "operationId": "5f701c6f291e60920062ec7b70afc3c7",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user's applications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Application"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/applications/{id}/status": {
            "patch": {
                "tags": [
                    "Applications"
                ],
                "summary": "Update application status (admin only)",
                "operationId": "ec30318084d4201cbbc31b04511465e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Application ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "reviewed",
                                            "accepted",
                                            "rejected"
                                        ]
                                    },
                                    "admin_notes": {
                                        "type": "string",
                                        "example": "Good candidate"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Application status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Application status updated successfully"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Application not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/applications/{id}": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get application detail",
                "operationId": "ebdb5654fcb10154e7462cfaf2af3042",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Application ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Application not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Application": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "job_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cover_letter": {
                        "type": "string",
                        "example": "I am interested..."
                    },
                    "cv_path": {
                        "type": "string",
                        "example": "cv/abc123.pdf"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "reviewed",
                            "accepted",
                            "rejected"
                        ]
                    },
                    "admin_notes": {
                        "type": "string",
                        "example": "Good candidate"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "job": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "title": {
                                "type": "string"
                            },
                            "company": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}